requirements.txt
Pyrogram
TgCrypto
py-tgcalls==0.5.5
python-dotenv
requests
aiohttp
aiofiles
asyncio
search_engine_parser
ffmpeg
Pillow
python-arq
future
psutil
wget
regex
dnspython
bs4
motor
python-telegram-bot==12.8
telegram
